(refactor): comment 코드 리팩토링 - history db를 참조하여 handler와 service의 리팩토링

+++ app/api/comment/resource/comment.go
@@ -2,8 +2,8 @@ package resource
 
 // comment
 type CommentResource struct {
-	Id       uint64 `json:"id"`
-	UserId   uint64 `json:"user_id"`
+	Id       int    `json:"id"`
+	UserId   int    `json:"user_id"`
 	UserName string `json:"nickname"`
 	Content  string `json:"content"`
 }
