+++ app/api/comment/resource/request.go
@@ -2,19 +2,19 @@ package resource
 
 // comment 요청값
 type GetAllCommentRequest struct {
-	Id uint64 `json:"id" query:"id"`
+	Id int `json:"id" query:"id"`
 }
 
 type CreateCommentRequest struct {
-	Id      uint64 `json:"id" query:"id"`
+	Id      int    `json:"id" query:"id"`
 	Content string `json:"content"`
 }
 
 type UpdateCommentRequest struct {
-	Id      uint64 `json:"id" query:"id"`
+	Id      int    `json:"id" query:"id"`
 	Content string `json:"content"`
 }
 
 type DeleteCommentRequest struct {
-	Id uint64 `json:"id" query:"id"`
+	Id int `json:"id" query:"id"`
 }
